Moved steinertree.py to approximation, moved the corresponding tests, changed set_up to setUp in test_steinertree.py
